+++ mycelium-ui/src/main.rs
@@ -1,56 +1,23 @@
 #![allow(non_snake_case)]
 
 mod api;
+mod components;
+
+use components::home::Home;
+use components::peers::Peers;
+use components::routes::Routes;
 
 use dioxus::prelude::*;
-use dioxus_free_icons::icons::fa_solid_icons::FaChevronLeft;
-use dioxus_free_icons::Icon;
-use human_bytes::human_bytes;
-use mycelium::peer_manager::PeerType;
-use std::collections::HashSet;
 use std::net::{IpAddr, Ipv4Addr, SocketAddr};
-use std::{cmp::Ordering, str::FromStr};
-use tracing::Level;
 
 const _: &str = manganis::mg!(file("assets/styles.css"));
 
-const DEFAULT_SERVER_ADDR: std::net::SocketAddr =
+const DEFAULT_SERVER_ADDR: SocketAddr =
     SocketAddr::new(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)), 8989);
-
-#[derive(Clone, Routable, Debug, PartialEq)]
-#[rustfmt::skip]
-pub enum Route {
-    #[layout(Layout)]
-        #[route("/")]
-        Home,
-        #[route("/peers")]
-        Peers,
-        #[route("/routes")]
-        Routes,
-    #[end_layout]
-    #[route("/:..route")]
-    PageNotFound { route: Vec<String> },
-}
-
-#[derive(Clone, PartialEq)]
-struct SearchState {
-    query: String,
-    column: String,
-}
-
-// Shared state structs below. We do not put both fields in one shared struct as
-// this can easily lead to a circular dependency when one field is being read whilst
-// the other is updated
-
-#[derive(Clone, PartialEq)]
-struct ServerAddress(SocketAddr);
-
-#[derive(Clone, PartialEq)]
-struct ServerConnected(bool);
 
 fn main() {
     // Init logger
-    dioxus_logger::init(Level::INFO).expect("failed to init logger");
+    dioxus_logger::init(tracing::Level::INFO).expect("failed to init logger");
     dioxus::launch(App);
 }
 
@@ -64,528 +31,34 @@ fn App() -> Element {
     }
 }
 
-#[component]
-fn Layout() -> Element {
-    let sidebar_collapsed = use_signal(|| false);
-
-    rsx! {
-        div { class: "app-container",
-            Header {}
-            div { class: "content-container",
-                Sidebar { collapsed: sidebar_collapsed }
-                main { class: if *sidebar_collapsed.read() { "main-content expanded" } else { "main-content" },
-                    Outlet::<Route> {}
-                }
-            }
-        }
-    }
+#[derive(Clone, Routable, Debug, PartialEq)]
+#[rustfmt::skip]
+pub enum Route {
+    #[layout(components::layout::Layout)]
+        #[route("/")]
+        Home {},
+        #[route("/peers")]
+        Peers,
+        #[route("/routes")]
+        Routes,
+    #[end_layout]
+    #[route("/:..route")]
+    PageNotFound { route: Vec<String> },
 }
-
-#[component]
-fn Header() -> Element {
-    let server_addr = use_context::<Signal<ServerAddress>>();
-    let fetched_node_info = use_resource(move || api::get_node_info(server_addr.read().0));
-    rsx! {
-        header {
-            h1 { "Mycelium Network Dashboard" }
-            div { class: "node-info",
-                { match &*fetched_node_info.read_unchecked() {
-                    Some(Ok(info)) => rsx! {
-                        span { "Subnet: {info.node_subnet}" }
-                        span { class: "separator", "|" }
-                        span { "Public Key: {info.node_pubkey}" }
-                    },
-                    Some(Err(_)) => rsx! { span { "Error loading node info" } },
-                    None => rsx! { span { "Loading node info..." } },
-                }}
-            }
-        }
-    }
-}
-
-#[component]
-fn Sidebar(collapsed: Signal<bool>) -> Element {
-    rsx! {
-        nav { class: if *collapsed.read() { "sidebar collapsed" } else { "sidebar" },
-            ul {
-                li { Link { to: Route::Home {}, "Home" } }
-                li { Link { to: Route::Peers {}, "Peers" } }
-                li { Link { to: Route::Routes {}, "Routes" } }
-            }
-        }
-        button { class: if *collapsed.read() { "toggle-sidebar collapsed" } else { "toggle-sidebar" },
-            onclick: {
-                let c = *collapsed.read();
-                move |_| collapsed.set(!c)
-            },
-            Icon {
-                icon: FaChevronLeft,
-            }
-        }
-    }
-}
-
-// This will fetch the node's subnet and public key and update the shared state
-// to set `connected` to true or false
-async fn fetch_node_info() -> Result<mycelium_api::Info, AppError> {
-    println!("Running fetch node info...");
-    let server_addr = use_context::<Signal<ServerAddress>>();
-    let mut server_connected = use_context::<Signal<ServerConnected>>();
-    let address = server_addr.read().0;
-
-    match api::get_node_info(address).await {
-        Ok(info) => {
-            server_connected.write().0 = true;
-
-            Ok(info)
-        }
-        Err(e) => {
-            server_connected.write().0 = true;
-            Err(AppError::from(e))
-        }
-    }
-}
-
-#[component]
-fn Home() -> Element {
-    println!("Running Home component");
-    let mut server_addr = use_context::<Signal<ServerAddress>>();
-    let mut new_address = use_signal(|| server_addr.read().0.to_string());
-    let mut node_info = use_resource(fetch_node_info);
-
-    let try_connect = move |_| {
-        if let Ok(addr) = SocketAddr::from_str(&new_address.read()) {
-            server_addr.write().0 = addr;
-            node_info.restart(); // This will trigger a new fetch_node_info() call
-        }
-    };
-
-    rsx! {
-        div { class: "home-container",
-            h2 { "Node information" }
-            div { class: "server-input",
-                input {
-                    placeholder: "Server address (e.g. 127.0.0.1:8989)",
-                    value: "{new_address}",
-                    oninput: move |evt| new_address.set(evt.value().clone()),
-                }
-                button { onclick: try_connect, "Connect" }
-            }
-            {match node_info.read().as_ref() {
-                Some(Ok(info)) => rsx! {
-                    p {
-                        "Node subnet: ",
-                        span { class: "bold", "{info.node_subnet}" }
-                    }
-                    p {
-                        "Node public key: ",
-                        span { class: "bold", "{info.node_pubkey}" }
-                    }
-                },
-                Some(Err(e)) => rsx! {
-                    p { class: "error", "Error: {e}" }
-                },
-                None => rsx! {
-                    p { "Enter a server address and click 'Connect' to fetch node information." }
-                }
-            }}
-        }
-    }
-}
-
-#[component]
-fn Peers() -> Element {
-    let server_addr = use_context::<Signal<ServerAddress>>().read().0;
-
-    // Use a signal to store the fetched peers
-    let mut peers = use_signal(|| Vec::new());
-
-    // Catch error if fetching peers fails
-    let mut error = use_signal(|| None::<String>);
-
-    // Use resource for initial fetch
-    let _initial_fetch = use_resource(move || async move {
-        match api::get_peers(server_addr).await {
-            Ok(fetched_peers) => {
-                peers.set(fetched_peers);
-                error.set(None);
-            }
-            Err(e) => {
-                println!("Error fetching peers: {}", e);
-                error.set(Some(format!(
-                    "An error has occurred while fetching peers: {}",
-                    e
-                )));
-            }
-        }
-    });
-
-    // Use coroutine for periodic updates
-    let _: Coroutine<()> = use_coroutine(|_rx: UnboundedReceiver<_>| {
-        // create clone of hook handles which can be moved into async closure
-        to_owned![peers, server_addr];
-        async move {
-            loop {
-                tokio::time::sleep(tokio::time::Duration::from_millis(500)).await;
-                match api::get_peers(server_addr).await {
-                    Ok(fetched_peers) => {
-                        peers.set(fetched_peers);
-                        error.set(None);
-                    }
-                    Err(e) => {
-                        println!("Error fetching peers: {}", e);
-                        error.set(Some(format!(
-                            "An error has occurred while fetching peers: {}",
-                            e
-                        )))
-                    }
-                }
-            }
-        }
-    });
-
-    // Render PeersTable with current peers data
-    rsx! {
-        if let Some(err) = error.read().as_ref() {
-            div { class: "error-message", "{err}" }
-        } else {
-            {PeersTable(peers)}
-        }
-    }
-}
-
-#[component]
-fn Routes() -> Element {
-    rsx! {
-        SelectedRoutesTable {}
-        FallbackRoutesTable {}
-    }
-}
-
-#[component]
-fn PageNotFound(route: Vec<String>) -> Element {
-    rsx! {
-        p { "Page not found"}
-    }
-}
-
-pub struct PeerTypeWrapper(pub mycelium::peer_manager::PeerType);
-impl Ord for PeerTypeWrapper {
-    fn cmp(&self, other: &Self) -> Ordering {
-        match (&self.0, &other.0) {
-            (PeerType::Static, PeerType::Static) => Ordering::Equal,
-            (PeerType::Static, _) => Ordering::Less,
-            (PeerType::LinkLocalDiscovery, PeerType::Static) => Ordering::Greater,
-            (PeerType::LinkLocalDiscovery, PeerType::LinkLocalDiscovery) => Ordering::Equal,
-            (PeerType::LinkLocalDiscovery, PeerType::Inbound) => Ordering::Less,
-            (PeerType::Inbound, PeerType::Inbound) => Ordering::Equal,
-            (PeerType::Inbound, _) => Ordering::Greater,
-        }
-    }
-}
-
-impl PartialOrd for PeerTypeWrapper {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl PartialEq for PeerTypeWrapper {
-    fn eq(&self, other: &Self) -> bool {
-        self.0 == other.0
-    }
-}
-
-impl Eq for PeerTypeWrapper {}
-
-#[derive(Clone)]
-enum SortDirection {
-    Ascending,
-    Descending,
-}
-
-fn sort_peers(
-    peers: &mut [mycelium::peer_manager::PeerStats],
-    column: &str,
-    direction: &SortDirection,
-) {
-    peers.sort_by(|a, b| {
-        let cmp = match column {
-            "Protocol" => a.endpoint.proto().cmp(&b.endpoint.proto()),
-            "Address" => a.endpoint.address().ip().cmp(&b.endpoint.address().ip()),
-            "Port" => a
-                .endpoint
-                .address()
-                .port()
-                .cmp(&b.endpoint.address().port()),
-            "Type" => PeerTypeWrapper(a.pt.clone()).cmp(&PeerTypeWrapper(b.pt.clone())),
-            "Connection State" => a.connection_state.cmp(&b.connection_state),
-            "Tx bytes" => a.tx_bytes.cmp(&b.tx_bytes),
-            "Rx bytes" => a.rx_bytes.cmp(&b.rx_bytes),
-            _ => Ordering::Equal,
-        };
-        match direction {
-            SortDirection::Ascending => cmp,
-            SortDirection::Descending => cmp.reverse(),
-        }
-    });
+//
+#[derive(Clone, PartialEq)]
+struct SearchState {
+    query: String,
+    column: String,
 }
 
 #[derive(Clone, PartialEq)]
-// Rows that have been expaneded out to show additional information
-struct ExpandedRows(HashSet<String>);
+struct ServerAddress(SocketAddr);
 
-fn PeersTable(peers: Signal<Vec<mycelium::peer_manager::PeerStats>>) -> Element {
-    let mut current_page = use_signal(|| 0);
-    let items_per_page = 20;
-    let mut sort_column = use_signal(|| "Protocol".to_string());
-    let mut sort_direction = use_signal(|| SortDirection::Ascending);
-    let peers_len = peers.len();
+#[derive(Clone, PartialEq)]
+struct ServerConnected(bool);
 
-    let mut change_page = move |delta: i32| {
-        let cur_page = *current_page.read() as i32;
-        current_page.set(
-            (cur_page + delta)
-                .max(0)
-                .min((peers_len - 1) as i32 / items_per_page as i32) as usize,
-        );
-    };
-
-    let mut expanded_rows = use_signal(|| ExpandedRows(HashSet::new()));
-    let mut toggle_row_expansion = move |peer_endpoint: String| {
-        let mut expanded = expanded_rows.write();
-        if expanded.0.contains(&peer_endpoint) {
-            expanded.0.remove(&peer_endpoint);
-        } else {
-            expanded.0.insert(peer_endpoint);
-        }
-    };
-
-    // Sorting
-    let mut sort_peers_signal = move |column: String| {
-        if column == *sort_column.read() {
-            let new_sort_direction = match *sort_direction.read() {
-                SortDirection::Ascending => SortDirection::Descending,
-                SortDirection::Descending => SortDirection::Ascending,
-            };
-            sort_direction.set(new_sort_direction);
-        } else {
-            sort_column.set(column);
-            sort_direction.set(SortDirection::Descending);
-        }
-        // When sorting, we should jump back to the first page
-        current_page.set(0);
-    };
-
-    let sorted_peers = use_memo(move || {
-        let mut sorted = peers.read().to_vec();
-        sort_peers(&mut sorted, &sort_column.read(), &sort_direction.read());
-        sorted
-    });
-
-    // Searching
-    let mut search_state = use_signal(|| SearchState {
-        query: String::new(),
-        column: "Protocol".to_string(),
-    });
-
-    let filtered_peers = use_memo(move || {
-        let query = search_state.read().query.to_lowercase();
-        let column = &search_state.read().column;
-        let sorted_peers = sorted_peers.read();
-        sorted_peers
-            .iter()
-            .filter(|peer| match column.as_str() {
-                "Protocol" => peer
-                    .endpoint
-                    .proto()
-                    .to_string()
-                    .to_lowercase()
-                    .contains(&query),
-                "Address" => peer
-                    .endpoint
-                    .address()
-                    .ip()
-                    .to_string()
-                    .to_lowercase()
-                    .contains(&query),
-                "Port" => peer
-                    .endpoint
-                    .address()
-                    .port()
-                    .to_string()
-                    .to_lowercase()
-                    .contains(&query),
-                "Type" => peer.pt.to_string().to_lowercase().contains(&query),
-                "Connection State" => peer
-                    .connection_state
-                    .to_string()
-                    .to_lowercase()
-                    .contains(&query),
-                "Tx bytes" => peer.tx_bytes.to_string().to_lowercase().contains(&query),
-                "Rx bytes" => peer.rx_bytes.to_string().to_lowercase().contains(&query),
-                _ => false,
-            })
-            .cloned()
-            .collect::<Vec<mycelium::peer_manager::PeerStats>>()
-    });
-
-    let peers_len = filtered_peers.read().len();
-
-    let start = current_page * items_per_page;
-    let end = (start + items_per_page).min(peers_len);
-    let current_peers = filtered_peers.read()[start..end].to_vec();
-
-    rsx! {
-        div { class: "peers-table",
-            h2 { "Peers" }
-            div { class: "search-container",
-                input {
-                    placeholder: "Search...",
-                    value: "{search_state.read().query}",
-                    oninput: move |evt| search_state.write().query.clone_from(&evt.value()),
-                }
-                select {
-                    value: "{search_state.read().column}",
-                    onchange: move |evt| search_state.write().column.clone_from(&evt.value()),
-                    option { value: "Protocol", "Protocol" }
-                    option { value: "Address", "Address" }
-                    option { value: "Port", "Port" }
-                    option { value: "Type", "Type" }
-                    option { value: "Connection State", "Connection State" }
-                    option { value: "Tx bytes", "Tx bytes" }
-                    option { value: "Rx bytes", "Rx bytes" }
-                }
-            }
-            div { class: "table-container",
-                table {
-                    thead {
-                        tr {
-                            th { class: "protocol-column",
-                                onclick: move |_| sort_peers_signal("Protocol".to_string()),
-                                "Protocol {get_sort_indicator(sort_column, sort_direction, \"Protocol\".to_string())}"
-                            }
-                            th { class: "address-column",
-                                onclick: move |_| sort_peers_signal("Address".to_string()),
-                                "Address {get_sort_indicator(sort_column, sort_direction, \"Address\".to_string())}"
-                            }
-                            th { class: "port-column",
-                                onclick: move |_| sort_peers_signal("Port".to_string()),
-                                "Port {get_sort_indicator(sort_column, sort_direction, \"Port\".to_string())}"
-                            }
-                            th { class: "type-column",
-                                onclick: move |_| sort_peers_signal("Type".to_string()),
-                                "Type {get_sort_indicator(sort_column, sort_direction, \"Type\".to_string())}"
-                            }
-                            th { class: "connection-state-column",
-                                onclick: move |_| sort_peers_signal("Connection State".to_string()),
-                                "Connection State {get_sort_indicator(sort_column, sort_direction, \"Connection State\".to_string())}"
-                            }
-                            th { class: "tx-bytes-column",
-                                onclick: move |_| sort_peers_signal("Tx bytes".to_string()),
-                                "Tx bytes {get_sort_indicator(sort_column, sort_direction, \"Tx bytes\".to_string())}"
-                            }
-                            th { class: "rx-bytes-column",
-                                onclick: move |_| sort_peers_signal("Rx bytes".to_string()),
-                                "Rx bytes {get_sort_indicator(sort_column, sort_direction, \"Rx bytes\".to_string())}"
-                            }
-                        }
-                    }
-                    tbody {
-                        for peer in current_peers.into_iter() {
-                            tr {
-                                onclick: move |_| toggle_row_expansion(peer.endpoint.to_string()),
-                                td { class: "protocol-column", "{peer.endpoint.proto()}" }
-                                td { class: "address-column", "{peer.endpoint.address().ip()}" }
-                                td { class: "port-column", "{peer.endpoint.address().port()}" }
-                                td { class: "type-column", "{peer.pt}" }
-                                td { class: "connection-state-column", "{peer.connection_state}" }
-                                td { class: "tx-bytes-column", "{human_bytes(peer.tx_bytes as f64)}" }
-                                td { class: "rx-bytes-column", "{human_bytes(peer.rx_bytes as f64)}" }
-                            }
-                            {
-                                let is_expanded = expanded_rows.read().0.contains(&peer.endpoint.to_string());
-                                if is_expanded {
-                                    rsx! {
-                                        ExpandedPeerRow {
-                                            peer: peer.clone(),
-                                            on_close: move |_| toggle_row_expansion(peer.endpoint.to_string())
-                                        }
-                                    }
-                                } else {
-                                    rsx! { }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            div { class: "pagination",
-                button {
-                    disabled: *current_page.read() == 0,
-                    onclick: move |_| change_page(-1),
-                    "Previous"
-                }
-                span { "Page {current_page + 1}" }
-                button {
-                    disabled: (current_page + 1) * items_per_page >= peers_len,
-                    onclick: move |_| change_page(1),
-                    "Next"
-                }
-            }
-        }
-    }
-}
-
-#[component]
-fn ExpandedPeerRow(peer: mycelium::peer_manager::PeerStats, on_close: EventHandler<()>) -> Element {
-    rsx! {
-        tr { class: "expanded-row",
-            td { colspan: "7",
-                div { class: "expanded-content",
-                    p { "Tx bytes: {human_bytes(peer.tx_bytes as f64)}" }
-                    p { "Rx bytes: {human_bytes(peer.rx_bytes as f64)}" }
-                    button { class: "close-button",
-                        onclick: move |_| on_close.call(()),
-                        "Close"
-                    }
-                }
-            }
-        }
-    }
-}
-
-#[component]
-fn SelectedRoutesTable() -> Element {
-    let server_addr = use_context::<Signal<ServerAddress>>();
-    let fetched_selected_routes =
-        use_resource(move || api::get_selected_routes(server_addr.read().0));
-
-    match &*fetched_selected_routes.read_unchecked() {
-        Some(Ok(routes)) => {
-            rsx! { { RoutesTable(routes.clone(), "Selected".to_string())  } }
-        }
-        Some(Err(e)) => rsx! { div { "An error has occurred while fetching selected routes: {e}" }},
-        None => rsx! { div { "Loading selected routes..." }},
-    }
-}
-
-#[component]
-fn FallbackRoutesTable() -> Element {
-    let server_addr = use_context::<Signal<ServerAddress>>();
-    let fetched_fallback_routes =
-        use_resource(move || api::get_fallback_routes(server_addr.read().0));
-
-    match &*fetched_fallback_routes.read_unchecked() {
-        Some(Ok(routes)) => {
-            rsx! { { RoutesTable(routes.clone(), "Fallback".to_string()) } }
-        }
-        Some(Err(e)) => rsx! { div { "An error has occurred while fetching fallback routes: {e}" }},
-        None => rsx! { div { "Loading fallback routes..." }},
-    }
-}
-
-fn get_sort_indicator(
+pub fn get_sort_indicator(
     sort_column: Signal<String>,
     sort_direction: Signal<SortDirection>,
     column: String,
@@ -600,155 +73,17 @@ fn get_sort_indicator(
     }
 }
 
-fn sort_routes(routes: &mut [mycelium_api::Route], column: &str, direction: &SortDirection) {
-    routes.sort_by(|a, b| {
-        let cmp = match column {
-            "Subnet" => a.subnet.cmp(&b.subnet),
-            "Next-hop" => a.next_hop.cmp(&b.next_hop),
-            "Metric" => a.metric.cmp(&b.metric),
-            "Seqno" => a.seqno.cmp(&b.seqno),
-            _ => Ordering::Equal,
-        };
-        match direction {
-            SortDirection::Ascending => cmp,
-            SortDirection::Descending => cmp.reverse(),
-        }
-    });
+#[component]
+fn PageNotFound(route: Vec<String>) -> Element {
+    rsx! {
+        p { "Page not found"}
+    }
 }
 
-fn RoutesTable(routes: Vec<mycelium_api::Route>, table_name: String) -> Element {
-    let mut current_page = use_signal(|| 0);
-    let items_per_page = 10;
-    let mut sort_column = use_signal(|| "Subnet".to_string());
-    let mut sort_direction = use_signal(|| SortDirection::Descending);
-    let routes_len = routes.len();
-
-    let mut change_page = move |delta: i32| {
-        let cur_page = *current_page.read() as i32;
-        current_page.set(
-            (cur_page + delta)
-                .max(0)
-                .min((routes_len - 1) as i32 / items_per_page as i32) as usize,
-        );
-    };
-
-    let mut sort_routes_signal = move |column: String| {
-        if column == *sort_column.read() {
-            let new_sort_direction = match *sort_direction.read() {
-                SortDirection::Ascending => SortDirection::Descending,
-                SortDirection::Descending => SortDirection::Ascending,
-            };
-            sort_direction.set(new_sort_direction);
-        } else {
-            sort_column.set(column);
-            sort_direction.set(SortDirection::Ascending);
-        }
-        // When sorting, we should jump back to the first page
-        current_page.set(0);
-    };
-
-    let sorted_routes = use_memo(move || {
-        let mut sorted = routes.clone();
-        sort_routes(&mut sorted, &sort_column.read(), &sort_direction.read());
-        sorted
-    });
-
-    // Searching
-    let mut search_state = use_signal(|| SearchState {
-        query: String::new(),
-        column: "Subnet".to_string(),
-    });
-
-    let filtered_routes = use_memo(move || {
-        let query = search_state.read().query.to_lowercase();
-        let column = &search_state.read().column;
-        sorted_routes
-            .read()
-            .iter()
-            .filter(|route| match column.as_str() {
-                "Subnet" => route.subnet.to_string().to_lowercase().contains(&query),
-                "Next-hop" => route.next_hop.to_string().to_lowercase().contains(&query),
-                "Metric" => route.metric.to_string().to_lowercase().contains(&query),
-                "Seqno" => route.seqno.to_string().to_lowercase().contains(&query),
-                _ => false,
-            })
-            .cloned()
-            .collect::<Vec<_>>()
-    });
-
-    let routes_len = filtered_routes.len();
-
-    let start = current_page * items_per_page;
-    let end = (start + items_per_page).min(routes_len);
-    let current_routes = &filtered_routes.read()[start..end];
-
-    rsx! {
-        div { class: "{table_name.to_lowercase()}-routes",
-            h2 { "{table_name} Routes" }
-            div { class: "search-container",
-                input {
-                    placeholder: "Search...",
-                    value: "{search_state.read().query}",
-                    oninput: move |evt| search_state.write().query.clone_from(&evt.value()),
-                }
-                select {
-                    value: "{search_state.read().column}",
-                    onchange: move |evt| search_state.write().column.clone_from(&evt.value()),
-                    option { value: "Subnet", "Subnet" }
-                    option { value: "Next-hop", "Next-hop" }
-                    option { value: "Metric", "Metric" }
-                    option { value: "Seqno", "Seqno" }
-                }
-            }
-            div { class: "table-container",
-                table {
-                    thead {
-                        tr {
-                            th { class: "subnet-column",
-                                onclick: move |_| sort_routes_signal("Subnet".to_string()),
-                                "Subnet {get_sort_indicator(sort_column, sort_direction, \"Subnet\".to_string())}"
-                            }
-                            th { class: "next-hop-column",
-                                onclick: move |_| sort_routes_signal("Next-hop".to_string()),
-                                "Next-hop {get_sort_indicator(sort_column, sort_direction, \"Next-hop\".to_string())}"
-                            }
-                            th { class: "metric-column",
-                                onclick: move |_| sort_routes_signal("Metric".to_string()),
-                                "Metric {get_sort_indicator(sort_column, sort_direction, \"Metric\".to_string())}"
-                            }
-                            th { class: "seqno_column",
-                                onclick: move |_| sort_routes_signal("Seqno".to_string()),
-                                "Seqno {get_sort_indicator(sort_column, sort_direction, \"Seqno\".to_string())}"
-                            }
-                        }
-                    }
-                    tbody {
-                        for route in current_routes {
-                            tr {
-                                td { class: "subnet-column", "{route.subnet}" }
-                                td { class: "next-hop-column", "{route.next_hop}" }
-                                td { class: "metric-column", "{route.metric}" }
-                                td { class: "seqno-column", "{route.seqno}" }
-                            }
-                        }
-                    }
-                }
-            }
-            div { class: "pagination",
-                button {
-                    disabled: *current_page.read() == 0,
-                    onclick: move |_| change_page(-1),
-                    "Previous"
-                }
-                span { "Page {current_page + 1}" }
-                button {
-                    disabled: (current_page + 1) * items_per_page >= routes_len,
-                    onclick: move |_| change_page(1),
-                    "Next"
-                }
-            }
-        }
-    }
+#[derive(Clone)]
+pub enum SortDirection {
+    Ascending,
+    Descending,
 }
 
 enum AppError {
